Populating shortest_path
Adding functionality to build the search algorithm for determining the shortest path between two people. Initialising a node, populating the frontier and then iterating through the frontier nodes until a solution is found.

Using the neighbors_for_person() function to explore where the next nodes are. Then check to see if any are the target and executing the solution code if this is True. Otherwise, populating the frontier with additional node options to iterate again.

If the frontier is empty, None is returned as expected.